PhotoEntry and PhotoFile are used by sample script, so move to shared
