+++ build.py
@@ -632,104 +632,6 @@ class RedPandaGraph:
         """All checks to ensure that the zoo dataset is good."""
         self.check_dataset_duplicate_ids(self.zoos)
 
-
-class PhotoEntry:
-    """
-    Represents all properties of a photo entry in a file.
-    Intended as a better way to track whether photos are new or not.
-    Git diffs are read line by line, and added to photo entities
-    based on their locator ID (<entity_id>.photo.<photo_id>).
-    
-    The UpdateFromCommits class uses hash tables of these by 
-    locator ID to make accurate counts of new photos, new entities, 
-    and new contributors.
-
-    Stores a reference to a filename cache, so we avoid rereading
-    configuration files every time a new locator is validated.
-    """
-    def __init__(self, filename, raw):
-        self.filename = filename
-        self.author_name = None
-        self.commitdate = None
-        self.entity_commitdate = None
-        self.entity_id = None
-        self.entity_type = None
-        self.photo_index = None
-        self.photo_uri = None
-        # Read in the entity details from the backing file just once
-        self._read_update_entity_id(raw)
-    
-    def entity_locator(self):
-        return self.entity_type + "." + self.entity_id
-
-    def photo_locator(self):
-        return self.entity_locator() + ".photo." + self.photo_index
-
-    def _read_update_entity_id(self, raw):
-        """
-        Open the config file and read its _id value.
-        Store the entity_type and entity_id for each photo.
-        
-        Read the raw line and return a corresponding entity ID to track
-        in one of the "media|panda|zoo" object tables. We're specifically
-        looking for updates matching the form: "photo.X: url".
-        """
-        key = raw.split(":")[0]
-        photo_uri = raw[len(key) + 1:].strip()
-        photo_index = key.split(".")[1]
-        config = configparser.ConfigParser()
-        self.filename = "./" + self.filename   # Standardize path
-        if not os.path.exists(self.filename):
-            # Fallback to filename number and path for entity
-            # This is a hack for when files are renamed
-            if self.filename.find(MEDIA_PATH) != -1:
-                self.entity_type = "media"
-            elif self.filename.find(PANDA_PATH) != -1:
-                self.entity_type = "panda"
-            elif self.filename.find(ZOO_PATH) != -1:
-                self.entity_type = "zoo"
-            # Consider whole path, and remove leading zeroes from id
-            self.entity_id = self.filename.split("/").pop().split("_")[0].lstrip("0")
-            # Other items will be entered as None
-            return
-        # Set all values based on entity and photo info        
-        config.read(self.filename, encoding='utf-8')
-        if self.filename.find(MEDIA_PATH) != -1:
-            entity = config.get("media", "_id")
-            self.entity_type = entity.split(".")[0]
-            self.entity_id = entity[len(self.entity_type) + 1:]
-            self.entity_commitdate = config.get("media", "commitdate")
-            self.author_name = config.get("media", key + ".author")
-            self.commitdate = config.get("media", key + ".commitdate")
-            self.photo_index = photo_index
-            self.photo_uri = photo_uri
-        elif self.filename.find(PANDA_PATH) != -1:
-            self.entity_type = "panda"
-            self.entity_id = config.get("panda", "_id")
-            self.entity_commitdate = config.get("panda", "commitdate")
-            self.author_name = config.get("panda", key + ".author")
-            self.commitdate = config.get("panda", key + ".commitdate")
-            self.photo_index = photo_index
-            self.photo_uri = photo_uri
-        elif self.filename.find(ZOO_PATH) != -1:
-            self.entity_type = "zoo"
-            self.entity_id = config.get("zoo", "_id")
-            self.entity_commitdate = config.get("zoo", "commitdate")
-            self.author_name = config.get("zoo", key + ".author")
-            self.commitdate = config.get("zoo", key + ".commitdate")
-            self.photo_index = photo_index
-            self.photo_uri = photo_uri
-        elif self.filename.find(WILD_PATH) != -1:
-            self.entity_type = "wild"
-            self.entity_id = config.get("wild", "_id")
-            self.entity_commitdate = config.get("wild", "commitdate")
-            self.author_name = config.get("wild", key + ".author")
-            self.commitdate = config.get("wild", key + ".commitdate")
-            self.photo_index = photo_index
-            self.photo_uri = photo_uri
-        else:
-            pass
-
 class UpdateFromCommits:
     """
     Take two Git commits in the repository, and process any additions
